Gruzii/requirements.txt
#    pip install -r requirements.txt

aiogram==3.21.0
python-telegram-bot==22.2
requests==2.32.5
python-dotenv==1.1.1